edit: AbcCHRONO > AbsYear and unit-delta

+++ src/TuttiFrutti/AbcChrono/Models/Hap.cs
@@ -4,7 +4,7 @@ class Hap : Universal, IHap
     public byte? DayNr { get; set; }
     public Month MonthName { get; set; } = Month.Undefined;
 
-    public uint Delta { get; set; } = 0;
+    public ushort Delta { get; set; } = 0;
 
     public ulong AbsYear { get => _absYear ?? NotSet.Throw(AbsYear); set => _absYear = value; }
     internal ulong? _absYear;
