+++ src/TuttiFrutti/AbcChrono/Models/IHap.cs
@@ -1,9 +1,9 @@
 ﻿namespace AbcChrono.Models;
-public interface IHap : IBasal, IConfidence<uint>
+public interface IHap : IBasal, IConfidence<ushort>
 {
     byte? DayNr { get; internal set; }
     Month MonthName { get; internal set; }
     ulong AbsYear { get; internal set; }
-    new uint Delta { get; internal set; }
+    new ushort Delta { get; internal set; }
 }
